Fix empty bq script

+++ testing/integration/tests/tf/scripts/empty-bq.sh
@@ -4,5 +4,7 @@ PROJECT_ID=$1
 DATASET_NAME=$2
 for i in $(bq ls $PROJECT_ID:$DATASET_NAME | sed 1,2d | awk "{print \$1}");
 do
-  bq rm -ft $PROJECT_ID:$DATASET_NAME.$i;
+  if  [[ ! -z "$i" ]]; then
+    bq rm -ft $PROJECT_ID:$DATASET_NAME.$i;
+  fi
 done
